linux-stats: containerize experiments
To make the experiments more reproducible, add a `Dockerfile` and
a `docker-compose.yaml` that implement a container. In this sense, the
only requirements for the user to run the experiments is to have
`docker` and `docker-compose` installed and a Linux mainline cloned.

Note that the README file should be updated with precise instructions
and any pertinent comment.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>